requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
langchain-google-genai==1.0.1
openai==1.14.0
faiss-cpu==1.8.0
pypdf==4.1.0
python-dotenv==1.0.1
tiktoken==0.6.0
